+++ packages/db/constants.js
@@ -1,0 +1,8 @@
+export const EMAIL_TYPE = {
+  User75PercentStorage: 'User75PercentStorage',
+  User80PercentStorage: 'User80PercentStorage',
+  User85PercentStorage: 'User85PercentStorage',
+  User90PercentStorage: 'User90PercentStorage',
+  User100PercentStorage: 'User100PercentStorage',
+  AdminStorageExceeded: 'AdminStorageExceeded'
+}