+++ packages/db/db-client-types.ts
@@ -28,6 +28,12 @@ export type UserOutput = {
   updated: definitions['user']['updated_at']
 }
 
+export type UserTagInput = {
+  tag: definitions['user_tag']['tag']
+  value: definitions['user_tag']['value']
+  reason?: definitions['user_tag']['reason']
+}
+
 // Auth key
 export interface CreateAuthKeyInput {
   name: definitions['auth_key']['name']
@@ -113,12 +119,7 @@ export type PinSyncRequestOutput = {
 }
 
 // Backup
-export type BackupOutput = {
-  _id: string
-  created: definitions['backup']['inserted_at']
-  url: definitions['backup']['url']
-  uploadId: definitions['backup']['upload_id']
-}
+export type BackupOutput = definitions['upload']['backup_urls']
 
 // Deal
 export type Deal = {
@@ -180,7 +181,7 @@ export interface CreateUploadInput {
     status: definitions['pin']['status']
     location: Location
   }>,
-  backupUrls: Array<definitions['backup']['url']>
+  backupUrls: definitions['upload']['backup_urls']
 }
 
 export type CreateUploadOutput = {
@@ -332,7 +333,33 @@ export type NameItem = {
   record: definitions['name']['record']
 }
 
-export type UsedStorage = {
+export type StorageUsedOutput = {
   uploaded: number,
-  pinned: number
+  psaPinned: number,
+  total: number
+}
+export type UserStorageUsedInput = {
+  fromPercent: number,
+  toPercent?: number
+}
+export type UserStorageUsedOutput = {
+  id: string,
+  name: string,
+  email: string,
+  storageQuota: number,
+  storageUsed: number,
+  percentStorageUsed: number
+}
+export type Email = {
+  type: definitions['email_history']['email_type']
+}
+export type EmailSentInput = {
+  userId: string,
+  emailType: string,
+  secondsSinceLastSent?: number
+}
+export type LogEmailSentInput = {
+  userId: string,
+  emailType: string,
+  messageId: string
 }